requirements.txt
flask
anthropic
gunicorn
httpx==0.27.2
pypdf
python-docx
